+++ Source/Runtime/Core/Public/Maths/ScalarBoundingBoxHelpers.h
@@ -26,38 +26,38 @@ T Volume(const TScalarBoxWExtent<T, 3>& aabb);
 //----------------------------------------------------------------------------
 //////////////////////////////////////////////////////////////////////////////
 //----------------------------------------------------------------------------
-template <typename T, size_t _Dim>
+template <typename T, u32 _Dim>
 TScalarVector<float, _Dim> LinearStep(const TScalarBoundingBox<T, _Dim>& aabb, const TScalarVector<T, _Dim>& value);
 //----------------------------------------------------------------------------
-template <typename U, typename T, size_t _Dim>
+template <typename U, typename T, u32 _Dim>
 TScalarVector<U, _Dim> QuantizeCeil(const TScalarBoundingBox<T, _Dim>& aabb, const TScalarVector<T, _Dim>& value);
 //----------------------------------------------------------------------------
-template <typename U, typename T, size_t _Dim>
+template <typename U, typename T, u32 _Dim>
 TScalarVector<U, _Dim> QuantizeFloor(const TScalarBoundingBox<T, _Dim>& aabb, const TScalarVector<T, _Dim>& value);
 //----------------------------------------------------------------------------
-template <typename U, typename T, size_t _Dim>
+template <typename U, typename T, u32 _Dim>
 TScalarVector<U, _Dim> QuantizeRound(const TScalarBoundingBox<T, _Dim>& aabb, const TScalarVector<T, _Dim>& value);
 //----------------------------------------------------------------------------
-template <typename U, typename T, size_t _Dim>
+template <typename U, typename T, u32 _Dim>
 TScalarVector<T, _Dim> Unquantize(const TScalarBoundingBox<T, _Dim>& aabb, const TScalarVector<U, _Dim>& quantized);
 //----------------------------------------------------------------------------
 //////////////////////////////////////////////////////////////////////////////
 //----------------------------------------------------------------------------
-template <typename T, size_t _Dim>
+template <typename T, u32 _Dim>
 TScalarBoundingBox<T, _Dim> MakeBoundingBox(const TMemoryView<TScalarVector<T, _Dim>>& points);
 //----------------------------------------------------------------------------
-template <typename T, size_t _Dim>
+template <typename T, u32 _Dim>
 TScalarBoundingBox<T, _Dim> MakeBoundingBox(const TMemoryView<const TScalarVector<T, _Dim>>& points);
 //----------------------------------------------------------------------------
-template <typename T, size_t _Dim>
+template <typename T, u32 _Dim>
 TScalarBoundingBox<T, _Dim> MakeBoundingBox(const TMemoryView<TScalarBoundingBox<T, _Dim>>& aabbs);
 //----------------------------------------------------------------------------
-template <typename T, size_t _Dim>
+template <typename T, u32 _Dim>
 TScalarBoundingBox<T, _Dim> MakeBoundingBox(const TMemoryView<const TScalarBoundingBox<T, _Dim>>& aabbs);
 //----------------------------------------------------------------------------
 //////////////////////////////////////////////////////////////////////////////
 //----------------------------------------------------------------------------
-template <typename T, size_t _Dim>
+template <typename T, u32 _Dim>
 void SwapEndiannessInPlace(TScalarBoundingBox<T, _Dim>* value) {
     SwapEndiannessInPlace(value->Min());
     SwapEndiannessInPlace(value->Max());
